requirement.txt
# Core Dependencies
Flask
flask-cors
flask[async]
Werkzeug

# AI and OCR Models
opencv-python
pytesseract
Pillow
numpy  # Compatible with all dependencies
scipy # Compatible with NumPy 1.23.0
sentence-transformers
transformers
moondream  # Avoids NumPy 2.x dependency

# NLP and Similarity Check
symspellpy
torch
torchvision
scikit-learn  # Requires NumPy <2.0

# PDF Processing
pdf2image
poppler-utils

# Miscellaneous
shutilwhich
asyncio
